requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
google-generativeai
pinecone-client
sentence-transformers